# appBackend/share/openapi.yaml
openapi: 3.0.0
info:
  title: 'Timesheet backend application'
  description: 'Api Timesheet backend application'
  version: 1.0.0
paths:
...
  /api/identity/users:
    get:
      tags:
        - Identity
      summary: 'Получить список пользователей'
      description: 'Возвращает список пользователей с пагинацией и фильтрацией.'
      operationId: get_identity_list_users
      parameters:
        -
          name: groupId
          in: query
          description: 'Фильтр по группе.'
          required: false
          schema:
            type: string
            format: uuid
        -
          name: roleId
          in: query
          description: 'Фильтр по роли.'
          required: false
          schema:
            type: string
            format: uuid
        -
          name: isActive
          in: query
          description: 'Фильтр по активности.'
          required: false
          schema:
            type: boolean
        -
          name: page
          in: query
          description: 'Номер страницы.'
          required: false
          schema:
            type: integer
            default: 1
        -
          name: perPage
          in: query
          description: 'Количество элементов на странице.'
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 'Список пользователей.'
          content:
            application/json:
              schema:
                $ref: App\Identity\Infrastructure\Dto\UserListResponseDto
    post:
      tags:
        - Identity
      summary: 'Создать пользователя'
      description: 'Создаёт нового пользователя.'
      operationId: post_identity_create_user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: App\Identity\Infrastructure\Dto\CreateUserRequestDto
      responses:
        '201':
          description: 'Пользователь успешно создан.'
        '400':
          description: 'Невалидные данные запроса.'
        '422':
          description: 'Нарушение бизнес-правил.'
...
  '/api/identity/groups/{id}':
    put:
      tags:
        - Identity
      summary: 'Обновить группу'
      description: 'Обновляет данные существующей группы.'
      operationId: put_identity_update_group
      parameters:
        -
          name: id
          in: path
          description: 'Идентификатор группы.'
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: App\Identity\Infrastructure\Dto\UpdateGroupRequestDto
      responses:
        '204':
          description: 'Группа успешно обновлена.'
        '404':
          description: 'Группа не найдена.'
        '422':
          description: 'Нарушение бизнес-правил.'
    delete:
      tags:
        - Identity
      summary: 'Удалить группу'
      description: 'Выполняет мягкое удаление группы.'
      operationId: delete_identity_delete_group
      parameters:
        -
          name: id
          in: path
          description: 'Идентификатор группы.'
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: 'Группа успешно удалена.'
        '404':
          description: 'Группа не найдена.'
        '422':
          description: 'Нарушение бизнес-правил.'
...
  /api/reporting/reports:
    get:
      tags:
        - Reporting
      summary: 'Получить список отчётов'
      description: 'Возвращает список отчётов с пагинацией и опциональной фильтрацией.'
      operationId: get_reporting_list_reports
      parameters:
        -
          name: createdBy
          in: query
          description: 'Фильтр по автору.'
          required: false
          schema:
            type: string
            format: uuid
        -
          name: periodFrom
          in: query
          description: 'Фильтр по началу периода (Y-m-d).'
          required: false
          schema:
            type: string
            format: date
        -
          name: periodTo
          in: query
          description: 'Фильтр по окончанию периода (Y-m-d).'
          required: false
          schema:
            type: string
            format: date
        -
          name: name
          in: query
          description: 'Фильтр по названию (поиск подстроки).'
          required: false
          schema:
            type: string
        -
          name: page
          in: query
          description: 'Номер страницы.'
          required: false
          schema:
            type: integer
            default: 1
        -
          name: perPage
          in: query
          description: 'Количество элементов на странице.'
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 'Список отчётов.'
          content:
            application/json:
              schema:
                $ref: App\Reporting\Infrastructure\Dto\ReportListResponseDto
    post:
      tags:
        - Reporting
      summary: 'Создать отчёт'
      description: 'Создаёт новый отчёт по заданным параметрам и фильтрам.'
      operationId: post_reporting_create_report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: App\Reporting\Infrastructure\Dto\CreateReportRequestDto
      responses:
        '201':
          description: 'Отчёт успешно создан.'
        '400':
          description: 'Невалидные данные запроса.'
        '422':
          description: 'Ошибка бизнес-логики.'
...
  '/api/reporting/reports/{id}':
    get:
      tags:
        - Reporting
      summary: 'Получить отчёт'
      description: 'Возвращает полные данные отчёта, включая строки агрегации.'
      operationId: get_reporting_get_report
      parameters:
        -
          name: id
          in: path
          description: 'Идентификатор отчёта.'
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 'Данные отчёта.'
          content:
            application/json:
              schema:
                $ref: App\Reporting\Infrastructure\Dto\ReportResponseDto
        '404':
          description: 'Отчёт не найден.'
...